Extract color resolver
