+++ packages/alfa-rules/src/common/dom/get-colors.ts
@@ -1,5 +1,6 @@
 /**
  * @internal
  */
+export * from "./get-colors/color";
 export * from "./get-colors/color-error";
 export * from "./get-colors/get-colors";
